move models to individual files
